requirements.txt
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
Naked==0.1.32
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.1
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
